+++ src/Catalog/Category/CategoryBridge.php
@@ -2,11 +2,16 @@
 
 namespace Shopgate\Shopware\Catalog\Category;
 
+use Shopgate\Shopware\Shopgate\Catalog\CategoryProductCollection;
 use Shopgate\Shopware\Storefront\ContextManager;
+use Shopgate\Shopware\System\Log\LoggerInterface;
 use Shopware\Core\Content\Category\CategoryCollection;
 use Shopware\Core\Content\Category\SalesChannel\AbstractCategoryListRoute;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\ContainsFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsAnyFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\RangeFilter;
 
 class CategoryBridge
@@ -14,7 +19,9 @@ class CategoryBridge
 
     public function __construct(
         private readonly AbstractCategoryListRoute $categoryListRoute,
-        private readonly ContextManager $contextManager
+        private readonly ContextManager $contextManager,
+        private readonly EntityRepository $categoryProductMapRepository,
+        private readonly LoggerInterface $logger
     ) {
     }
 
@@ -23,7 +30,7 @@ class CategoryBridge
         return $this->contextManager->getSalesContext()->getSalesChannel()->getNavigationCategoryId();
     }
 
-    public function getChildCategories(string $parentId): CategoryCollection
+    public function getChildCategories(string $parentId, ?int $offset = null): CategoryCollection
     {
         $criteria = (new Criteria())
             ->addAssociation('media')
@@ -34,12 +41,32 @@ class CategoryBridge
                     RangeFilter::GT => 1,
                     RangeFilter::LTE => 99,
                 ])
-            );
+            )
+            ->setOffset($offset);
         $criteria->setTitle('shopgate::category::parent-id');
         $list = $this->categoryListRoute->load($criteria, $this->contextManager->getSalesContext())->getCategories();
         $tree = $this->buildTree($parentId, $list);
 
         return $this->flattenTree($tree, new CategoryCollection());
+    }
+
+    /**
+     * @param string[] $uids
+     * @return CategoryProductCollection
+     */
+    public function getCategoryProductMap(array $uids = []): CategoryProductCollection
+    {
+        $channel = $this->contextManager->getSalesContext();
+        $criteria = (new Criteria());
+        $criteria->addFilter(new EqualsFilter('salesChannelId', $channel->getSalesChannelId()));
+        if ($uids) {
+            $criteria->addFilter(new EqualsAnyFilter('productId', $uids));
+        }
+
+        $entities = $this->categoryProductMapRepository->search($criteria, $channel->getContext())->getEntities();
+        $entities->count() === 0 && $this->logger->debug('No category/product mapping entities found in index');
+
+        return $entities;
     }
 
     private function buildTree(?string $parentId, CategoryCollection $categories): CategoryCollection