+++ src/Catalog/Product/Sort/SortTree.php
@@ -25,106 +25,15 @@ class SortTree
     public const CACHE_KEY = 'shopgate.category.sort';
 
     public function __construct(
-        private readonly TagAwareAdapterInterface $cache,
         private readonly ContextManager $contextManager,
-        private readonly CategoryBridge $categoryBridge,
-        private readonly ConfigBridge $configReader,
-        private readonly AbstractProductListRoute $listRoute,
         private readonly AbstractProductListingRoute $listingRoute,
-        private readonly LoggerInterface $logger
     ) {
-    }
-
-    /**
-     * To enable cache you will need to go "prod" mode & enable HTTP Cache
-     * @throws InvalidArgumentException|CacheException
-     */
-    public function getSortTree(string $rootCategoryId): array
-    {
-        $tree = $this->cache->getItem(self::CACHE_KEY . '.' . $rootCategoryId);
-        try {
-            if ($tree->isHit() && $tree->get()) {
-                return CacheCompressor::uncompress($tree);
-            }
-        } catch (Throwable $e) {
-            $this->logger->error($e->getMessage());
-        }
-
-        $this->logger->debug('Building new sort order cache');
-
-        if ($this->configReader->get(ConfigBridge::SYSTEM_CONFIG_IGNORE_SORT_ORDER)) {
-            $build = $this->buildWithoutPosition($rootCategoryId);
-        } else {
-            $build = $this->buildWithPosition($rootCategoryId);
-        }
-
-        $tree = CacheCompressor::compress($tree, $build);
-        $tree->tag([self::CACHE_KEY]);
-        $this->cache->save($tree);
-
-        return $build;
-    }
-
-    /**
-     * @param string $rootCategoryId - provide category id to build from
-     * @return array - ['categoryId' => ['productId' => sortNumber]]
-     */
-    private function buildWithPosition(string $rootCategoryId): array
-    {
-        $tree = [];
-        $categories = $this->categoryBridge->getChildCategories($rootCategoryId);
-        foreach ($categories as $category) {
-            $products = $this->getAllCategoryProducts($category);
-            $maxProducts = $products->count();
-            $i = 0;
-            foreach ($products as $product) {
-                $tree[$product->getParentId() ?: $product->getId()][] = [
-                    'categoryId' => $category->getId(),
-                    'position' => $maxProducts - $i++
-                ];
-            }
-        }
-
-        return $tree;
-    }
-
-    /**
-     * @param string $rootCategoryId - provide category id to build from
-     * @return array - ['categoryId' => ['productId']]
-     */
-    private function buildWithoutPosition(string $rootCategoryId): array
-    {
-        $tree = [];
-        $categories = $this->categoryBridge->getChildCategories($rootCategoryId);
-        $products = $this->getAllProducts();
-
-        foreach ($products as $product) {
-
-            foreach ($categories as $category) {
-
-                $identifier = $product->getParentId() ?: $product->getId();
-                if (array_key_exists($identifier, $tree)
-                    && in_array($category->getId(), array_column($tree[$identifier], 'categoryId'), true)) {
-                    continue;
-                }
-
-                if (($product->getCategoryTree() && in_array($category->getId(), $product->getCategoryTree(), true)
-                        || $product->getStreamIds() && in_array($category->getProductStreamId(),
-                            $product->getStreamIds(), true))
-                    || $this->hasChildren($category, $product)) {
-
-                    $tree[$identifier][] = ['categoryId' => $category->getId()];
-                }
-            }
-        }
-
-        return $tree;
     }
 
     /**
      * Loops through all products for every category out there. Expensive stuff!
      */
-    private function getAllCategoryProducts(CategoryEntity $category): ProductCollection
+    public function getAllCategoryProducts(CategoryEntity $category): ProductCollection
     {
         $list = new ProductCollection();
         $page = 1;
@@ -152,17 +61,6 @@ class SortTree
     }
 
     /**
-     * Loops through all products for every category out there. Expensive stuff!
-     */
-    private function getAllProducts(): ProductCollection
-    {
-        $criteria = new Criteria();
-        $criteria->setTitle('shopgate::product::all');
-
-        return $this->listRoute->load($criteria, $this->contextManager->getSalesContext())->getProducts();
-    }
-
-    /**
      * Retrieves the default key to sort the category by
      *
      * @param CategoryEntity $category
@@ -181,22 +79,5 @@ class SortTree
         }
 
         return null;
-    }
-
-    private function hasChildren(CategoryEntity $category, SalesChannelProductEntity $product): bool
-    {
-        $hasChild = false;
-        /** @var CategoryEntity $child */
-        foreach ($category->getChildren() as $child) {
-
-            if ($child->getChildren() && $child->getChildren()->count()) {
-                $this->hasChildren($child, $product);
-            }
-            $hasChild =
-                ($product->getCategoryTree() && in_array($child->getId(), $product->getCategoryTree(), true))
-                || ($product->getStreamIds() && in_array($child->getProductStreamId(), $product->getStreamIds(), true));
-        }
-
-        return $hasChild;
     }
 }