+++ src/Shopgate/Catalog/CategoryProductIndexingMessage.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace Shopgate\Shopware\Shopgate\Catalog;
+
+use Shopware\Core\Framework\DataAbstractionLayer\Indexing\EntityIndexingMessage;
+
+class CategoryProductIndexingMessage extends EntityIndexingMessage
+{
+}